+++ src/Controller/SecurityController.php
@@ -4,12 +4,15 @@ namespace App\Controller;
 
 use App\Entity\User;
 use App\Enum\PlatformMode;
+use App\Enum\TwoFAType;
+use App\Enum\UserTwoFactorAuthenticationStatus;
 use App\Form\LoginFormType;
 use App\Repository\SettingRepository;
 use App\Repository\UserRepository;
 use App\Service\GetSettings;
 use Doctrine\ORM\NonUniqueResultException;
 use LogicException;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -21,15 +24,15 @@ class SecurityController extends AbstractController
 {
     /**
      * SiteController constructor.
-     *
      * @param UserRepository $userRepository The repository for accessing user data.
      * @param SettingRepository $settingRepository The setting repository is used to create the getSettings function.
      * @param GetSettings $getSettings The instance of GetSettings class.
+     *  of the user account
      */
     public function __construct(
         private readonly UserRepository $userRepository,
         private readonly SettingRepository $settingRepository,
-        private readonly GetSettings $getSettings
+        private readonly GetSettings $getSettings,
     ) {
     }
 
@@ -45,23 +48,91 @@ class SecurityController extends AbstractController
         $user_sigin = new User();
         $form = $this->createForm(LoginFormType::class, $user_sigin);
         $form->handleRequest($request);
+        $user = $this->getUser();
 
         // Check if the user is already logged in and redirect them accordingly
-        if ($this->getUser() instanceof \Symfony\Component\Security\Core\User\UserInterface) {
-            if ($type === 'admin') {
-                if ($this->isGranted('ROLE_ADMIN')) {
-                    $session = $request->getSession();
-                    $session->set('session_admin', true);
-                    return $this->redirectToRoute('admin_page');
-                }
-                $this->addFlash('error', 'Wrong credentials');
-                return $this->redirectToRoute('saml_logout');
+        if ($user instanceof User) {
+            if ($this->isGranted('ROLE_ADMIN')) {
+                $session = $request->getSession();
+                $session->set('session_admin', true);
+                return $this->redirectToRoute('admin_page');
             }
             $platformMode = $data['PLATFORM_MODE']['value'];
             if ($platformMode === PlatformMode::DEMO->value) {
                 return $this->redirectToRoute('saml_logout');
             }
-
+            // Get 2fa status on platform
+            $twoFAplatformStatus = $this->settingRepository->findOneBy(['name' => 'TWO_FACTOR_AUTH_STATUS']);
+            if ($twoFAplatformStatus) {
+                // Check 2fa status on platform, after that we need to check user status to decide what case we have
+                if ($twoFAplatformStatus->getValue() === TwoFAType::NOT_ENFORCED->value) {
+                    if (
+                        $user->getTwoFAType() ===
+                        UserTwoFactorAuthenticationStatus::DISABLED->value ||
+                        !$user->getTwoFAType() instanceof UserTwoFactorAuthenticationStatus
+                    ) {
+                        return $this->redirectToRoute('app_landing');
+                    }
+                    if (
+                        $user->getTwoFAType() ===
+                        UserTwoFactorAuthenticationStatus::SMS->value
+                    ) {
+                        return $this->redirectToRoute('app_verify2FA_local');
+                    }
+                    if (
+                        $user->getTwoFAType() ===
+                        UserTwoFactorAuthenticationStatus::APP->value
+                    ) {
+                        return $this->redirectToRoute('app_verify2FA_app');
+                    }
+                    return $this->redirectToRoute('app_landing');
+                }
+                if ($twoFAplatformStatus->getValue() === TwoFAType::ENFORCED_FOR_LOCAL->value) {
+                    if (
+                        $user->getTwoFAType() ===
+                        UserTwoFactorAuthenticationStatus::DISABLED->value ||
+                        !$user->getTwoFAType() instanceof UserTwoFactorAuthenticationStatus
+                    ) {
+                        return $this->redirectToRoute('app_enable2FA');
+                    }
+                    if (
+                        $user->getTwoFAType() ===
+                        UserTwoFactorAuthenticationStatus::SMS->value
+                    ) {
+                        return $this->redirectToRoute('app_verify2FA_local');
+                    }
+                    if (
+                        $user->getTwoFAType() ===
+                        UserTwoFactorAuthenticationStatus::APP->value
+                    ) {
+                        return $this->redirectToRoute('app_verify2FA_app');
+                    }
+                    return $this->redirectToRoute('app_verify2FA_local');
+                }
+                if ($twoFAplatformStatus->getValue() === TwoFAType::ENFORCED_FOR_ALL->value) {
+                    if (
+                        $user->getTwoFAType() ===
+                        UserTwoFactorAuthenticationStatus::DISABLED->value ||
+                        !$user->getTwoFAType() instanceof UserTwoFactorAuthenticationStatus
+                    ) {
+                        return $this->redirectToRoute('app_enable2FA');
+                    }
+                    if (
+                        $user->getTwoFAType() ===
+                        UserTwoFactorAuthenticationStatus::SMS->value
+                    ) {
+                        return $this->redirectToRoute('app_verify2FA_local');
+                    }
+                    if (
+                        $user->getTwoFAType() ===
+                        UserTwoFactorAuthenticationStatus::APP->value
+                    ) {
+                        return $this->redirectToRoute('app_verify2FA_app');
+                    }
+                    return $this->redirectToRoute('app_enable2FA');
+                }
+                return $this->redirectToRoute('app_landing');
+            }
             return $this->redirectToRoute('app_landing');
         }
 
@@ -87,7 +158,7 @@ class SecurityController extends AbstractController
             $this->addFlash('error', $error->getMessage());
         }
 
-        if ($type === "admin") {
+        if ($this->isGranted('ROLE_ADMIN')) {
             return $this->render('admin/login_admin_landing.html.twig', [
                 'last_username' => $lastUsername,
                 'error' => $error,